Implement tokenize_input()
While awaring that the input from users are extensively various,
I assumed that the input should be separated by only comma. This
is merely obedience to this project's specification.

For improvement, you need to trim white spaces at head and tail.
Also, without this convenient assumption, you need to check if
each of those given arguments are valid.

Resolves: #2