Add register variables
A variable of struct register type has all register variables.

Note that to mimic how CPU works better, it is more natural to use an array of
registers accessing each register by using register number as an index. However,
there are some registers that are not used in this project such as F and SW.
I decied to use structure because it doesn't waste any memory.

Resolves: #11